# public/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "🚗 Auto-École API - Authentification Authentik",
        "description": "API complète pour la gestion de l'authentification avec Authentik comme IAM central. Cette API permet l'inscription, la connexion, la gestion des utilisateurs et des tokens avec un système de contournement du Password Grant Flow.",
        "contact": {
            "name": "Support Auto-École API",
            "url": "https://auto-ecole-api.com",
            "email": "support@auto-ecole-api.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur de développement local"
        },
        {
            "url": "https://9c8r7bbvybn.preview.infomaniak.website",
            "description": "Serveur de production Infomaniak"
        },
        {
            "url": "http://localhost:8000",
            "description": "Serveur de développement local"
        },
        {
            "url": "https://9c8r7bbvybn.preview.infomaniak.website",
            "description": "Serveur de production Infomaniak"
        },
        {
            "url": "http://localhost:8000",
            "description": "Serveur de développement local"
        },
        {
            "url": "https://9c8r7bbvybn.preview.infomaniak.website",
...
        },
        {
...
        },
...
    "paths": {
...
        "/api/documents": {
            "get": {
                "tags": [
                    "📄 Documents"
                ],
                "summary": "📋 Liste des documents",
                "operationId": "getDocuments",
                "parameters": [
                    {
                        "name": "dossier_id",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "valide",
                        "in": "query",
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "✅ Liste des documents"
                    }
                }
            },
            "post": {
                "tags": [
                    "📄 Documents"
                ],
                "summary": "➕ Créer un document",
                "operationId": "storeDocument",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "dossier_id",
                                    "type_document_id",
                                    "nom_fichier",
                                    "chemin_fichier",
                                    "type_mime",
                                    "taille_fichier"
                                ],
                                "properties": {
                                    "dossier_id": {
                                        "type": "string",
                                        "format": "uuid"
                                    },
                                    "type_document_id": {
                                        "type": "string",
                                        "format": "uuid"
                                    },
                                    "nom_fichier": {
                                        "type": "string",
                                        "example": "carte_identite.pdf"
                                    },
                                    "chemin_fichier": {
                                        "type": "string",
                                        "example": "/uploads/documents/carte_identite.pdf"
                                    },
                                    "type_mime": {
                                        "type": "string",
                                        "example": "application/pdf"
                                    },
                                    "taille_fichier": {
                                        "type": "integer",
                                        "example": 1024000
                                    },
                                    "valide": {
                                        "type": "boolean",
                                        "example": false
                                    },
                                    "commentaires": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "✅ Document créé"
                    }
                }
            }
        },
        "/api/documents/{id}": {
            "get": {
                "tags": [
                    "📄 Documents"
                ],
                "summary": "🔍 Détails d'un document",
                "operationId": "getDocument",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "✅ Détails du document"
                    }
                }
            },
            "put": {
                "tags": [
                    "📄 Documents"
                ],
                "summary": "✏️ Mettre à jour un document",
                "operationId": "updateDocument",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "valide": {
                                        "type": "boolean"
                                    },
                                    "commentaires": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "✅ Document mis à jour"
                    }
                }
            },
            "delete": {
                "tags": [
                    "📄 Documents"
                ],
                "summary": "🗑️ Supprimer un document",
                "operationId": "deleteDocument",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "✅ Document supprimé"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "description": "Token d'accès JWT obtenu via l'authentification",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "bearer_token": {
                "type": "http",
                "description": "Token d'authentification JWT obtenu via /api/auth/login-direct",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
}